Added better descriptions for functions, improved reliability, repackaged objects.
Former-commit-id: f87e4dd47d11a7c26d3781fdc8594153a7794612

+++ src/nl/b3p/kaartenbalie/core/server/reporting/control/RequestReporting.java
@@ -18,10 +18,10 @@ import java.util.Map;
 import javax.persistence.EntityManager;
 import javax.persistence.EntityTransaction;
 import nl.b3p.kaartenbalie.core.server.persistence.ManagedPersistence;
-import nl.b3p.kaartenbalie.core.server.reporting.domain.ClientRequest;
+import nl.b3p.kaartenbalie.core.server.reporting.domain.requests.ClientRequest;
 import nl.b3p.kaartenbalie.core.server.User;
-import nl.b3p.kaartenbalie.core.server.reporting.domain.RequestOperation;
-import nl.b3p.kaartenbalie.core.server.reporting.domain.ServiceProviderRequest;
+import nl.b3p.kaartenbalie.core.server.reporting.domain.operations.RequestOperation;
+import nl.b3p.kaartenbalie.core.server.reporting.domain.requests.ServiceProviderRequest;
 import javax.xml.parsers.DocumentBuilder;
 import javax.xml.parsers.DocumentBuilderFactory;
 
@@ -33,8 +33,15 @@ public class RequestReporting {
     private static List usRequestOperation;
     private long reportStopwatch = 0;
     private long lapTimer;
-    private static boolean outputLaptimer = false;
     
+    /*
+     * Basically use this boolean to enable or disable the logging mechanism.
+     */
+    private static boolean enableReporting = false;
+    
+    /* These are static sets used for the reflectionfunction. It makes sure that it isnt possible to the Id and client
+     * requestvalues for the ServiceProviderRequests and RequestOperations.
+     */
     static {
         usServiceProviderRequest = new ArrayList();
         usServiceProviderRequest.add(new String("setClientRequest"));
@@ -45,19 +52,28 @@ public class RequestReporting {
         usRequestOperation.add(new String("setId"));
     }
     
-    public RequestReporting() {
+    private RequestReporting() {
         em = ManagedPersistence.createEntityManager();
     }
+    
+    /*
+     * This should be the default call to create a requestReporting object.
+     */
     public RequestReporting(User user) {
         this();
         if (user != null) {
-            
             this.user = (User) em.find(User.class, user.getId());
         }
-        
     }
     
+    /*
+     * For testing purposes only. This function will be removed in the future
+     */
     public void clean() {
+        if (transactionActive()) {
+            throw new Error("Cannot run the clean command when a reporting is in progress.. ");
+        }
+        
         EntityTransaction tx = em.getTransaction();
         tx.begin();
         em.createQuery("DELETE FROM ServiceProviderRequest").executeUpdate();
@@ -67,9 +83,14 @@ public class RequestReporting {
         
     }
     
+    /*
+     * This is your init for to create a new RequestReport. Call it whenever you feel like starting a new
+     * clientreport! You cannot make multiple reports at once though.
+     */
     public void startClientRequest(String clientRequestURI, int bytesReceivedFromUser) {
+        if (!enableReporting) return;
         EntityTransaction tx = em.getTransaction();
-        if (tx.isActive() || clientRequest != null) {
+        if (transactionActive()) {
             throw new Error("Cannot start a new clientRequest without ending one first.");
         }
         tx.begin();
@@ -77,25 +98,60 @@ public class RequestReporting {
             clientRequest = new ClientRequest();
             clientRequest.setClientRequestURI(clientRequestURI);
             clientRequest.setBytesReceivedFromUser(new Integer(bytesReceivedFromUser));
-//            em.persist(clientRequest);
+            em.persist(clientRequest);
         } catch (Exception e) {
             e.printStackTrace();
         }
     }
     
+    /*
+     * The function addServiceProvider request makes use of the reflection method described below. It is used to log
+     * every call to a ServiceProvider. It should contain a parameterMap that matches the setters of the specified
+     * sprClass.
+     */
     public void addServiceProviderRequest(Class sprClass, Map parameterMap){
+        if (!enableReporting) return;
+        
+        if (!transactionActive()) {
+            throw new Error("Please start a clientrequest before attempting to log ServiceProviderRequests");
+        }
         Map overriddenParameters = new HashMap();
         overriddenParameters.put("setClientRequest", clientRequest);
         Object reflectedObject = RequestReporting.createPOJOByReflection(ServiceProviderRequest.class, sprClass, parameterMap, overriddenParameters, usServiceProviderRequest);
-//        em.persist(reflectedObject);
+        em.persist(reflectedObject);
     }
     
+    /*
+     * The function addRequestOperation request makes use of the reflection method described below. It is used to log
+     * every call to a RequestOperation. It should contain a parameterMap that matches the setters of the specified
+     * rqoClass.
+     */
     public void addRequestOperation(Class rqoClass, Map parameterMap){
+        if (!enableReporting) return;
+        if (!transactionActive()) {
+            throw new Error("Please start a clientrequest before attempting to log RequestOperations");
+        }
+        
         Map overriddenParameters = new HashMap();
         overriddenParameters.put("setClientRequest", clientRequest);
         Object reflectedObject = RequestReporting.createPOJOByReflection(RequestOperation.class, rqoClass, parameterMap, overriddenParameters, usRequestOperation);
-//        em.persist(reflectedObject);
+        em.persist(reflectedObject);
     }
+    
+    /*
+     * This is a small function which will check wether a transaction is in a valid state in combination with a
+     * clientRequest. It is used in different perspectives all throughout this class.
+     */
+    private boolean transactionActive() {
+        EntityTransaction tx = em.getTransaction();
+        if (tx.isActive() && clientRequest == null) {
+            throw new Error("Transaction is active, but there is no clientrequest..");
+        } else if (!tx.isActive() && clientRequest != null) {
+            throw new Error("There is a clientRequest, but the transaction is not active!");
+        }
+        return (tx.isActive() && clientRequest != null);
+    }
+    
     
     
     /*
@@ -157,7 +213,6 @@ public class RequestReporting {
                         (unsupportedMethods == null || (unsupportedMethods != null && !unsupportedMethods.contains(method.getName())))) {
                     String methodName = method.getName();
                     methodName = methodName.substring(3,methodName.length());
-                    
                     Class[] parameterTypes = method.getParameterTypes();
                     validMessage += methodName + "(";
                     for (int j = 0; j < parameterTypes.length; j++) {
@@ -174,10 +229,15 @@ public class RequestReporting {
         return null;
     }
     
+    /*
+     * This is your final statement in logging call. It will clean up your clientRequest and commit all the logged calls to
+     * the database. After this you can restart your clientRequest without creating a new RequestReporting.
+     */
     public void endClientRequest(int bytesSendToUser, long totalResponseTime) {
+        if (!enableReporting) return;
         EntityTransaction tx = em.getTransaction();
         
-        if (!tx.isActive() || clientRequest == null) {
+        if (!transactionActive()) {
             throw new Error("Cannot end a clientRequest without starting it first.");
         }
         
@@ -192,20 +252,14 @@ public class RequestReporting {
         }
     }
     
-    
-    
-    
-    public static void main(String [ ] args) throws Exception {
-        
-        
-        
-        
+    public static void main(String [] args) throws Exception {
+        EntityManager em = ManagedPersistence.createEntityManager();
+        User user = (User) em.find(User.class, new Integer(1) );
+        RequestReporting rr = new RequestReporting(user);
+        rr.clean();
+        TestThread tt = new TestThread();
+        tt.start();
     }
-
-    
-    
-    
-
